backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
psycopg[binary,pool]>=3.2.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
starlette>=0.41.0
httpx>=0.27.0
PyJWT>=2.8.0
requests>=2.31.0
openai>=1.0.0
boto3>=1.28.0
loguru>=0.7.0
